hookup standalone react component
