+++ common/components/chrome/MainHeader.jsx
@@ -44,18 +44,11 @@ class MainHeader extends React.PureComponent<{||}> {
       )
       : (
         <span>
-          <a href="#" onClick={this._onLogInClick}>Log In</a> |{' '}
+          <a href="/login_react">Log In</a> |{' '}
           <a href="/signup">Sign Up</a> |{' '}
           <a href="/password_reset">Forgot Password</a>
         </span>
       );
-  }
-
-  _onLogInClick(): void {
-    NavigationDispatcher.dispatch({
-      type: 'SET_SECTION',
-      section: Section.SignIn,
-    });
   }
 
   _renderLinks(): React$Node {
