+++ common/components/enums/Section.js
@@ -5,7 +5,6 @@ const Section = {
   MyProjects: 'MyProjects',
   Profile: 'Profile',
   Inbox: 'Inbox',
-  SignIn: 'SignIn',
 };
 
 export type SectionType = $Keys<typeof Section>;
